+++ frontend/src/FirstRun/AuthenticationRequiredModalContentConnector.js
@@ -3,6 +3,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
 import { clearPendingChanges } from 'Store/Actions/baseActions';
+import { fetchCurrentUser } from 'Store/Actions/currentUserActions';
 import { fetchGeneralSettings, saveGeneralSettings, setGeneralSettingsValue } from 'Store/Actions/settingsActions';
 import { fetchStatus } from 'Store/Actions/systemActions';
 import createSettingsSectionSelector from 'Store/Selectors/createSettingsSectionSelector';
@@ -26,16 +27,32 @@ const mapDispatchToProps = {
   dispatchSetGeneralSettingsValue: setGeneralSettingsValue,
   dispatchSaveGeneralSettings: saveGeneralSettings,
   dispatchFetchGeneralSettings: fetchGeneralSettings,
-  dispatchFetchStatus: fetchStatus
+  dispatchFetchStatus: fetchStatus,
+  dispatchFetchCurrentUser: fetchCurrentUser
 };
 
 class AuthenticationRequiredModalContentConnector extends Component {
+  constructor(props, context) {
+    super(props, context);
+
+    this._pendingLogin = null;
+  }
 
   //
   // Lifecycle
 
   componentDidMount() {
     this.props.dispatchFetchGeneralSettings();
+  }
+
+  componentDidUpdate(prevProps) {
+    if (prevProps.isSaving && !this.props.isSaving) {
+      if (!this.props.saveError && this._pendingLogin) {
+        this.login(this._pendingLogin);
+      }
+
+      this._pendingLogin = null;
+    }
   }
 
   componentWillUnmount() {
@@ -50,7 +67,46 @@ class AuthenticationRequiredModalContentConnector extends Component {
   };
 
   onSavePress = () => {
+    const settings = this.props.settings || {};
+    const username = settings.username?.value ?? '';
+    const password = settings.password?.value ?? '';
+
+    this._pendingLogin = {
+      username,
+      password
+    };
+
     this.props.dispatchSaveGeneralSettings();
+  };
+
+  login = async ({ username, password }) => {
+    if (!username || !password) {
+      return;
+    }
+
+    const loginUrl = `${window.Readarr.urlBase}/login`;
+    const body = new URLSearchParams({
+      username,
+      password,
+      rememberMe: 'true'
+    });
+
+    try {
+      const response = await fetch(loginUrl, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/x-www-form-urlencoded'
+        },
+        body,
+        credentials: 'same-origin'
+      });
+
+      if (response.ok && !response.url.includes('loginFailed=true')) {
+        this.props.dispatchFetchCurrentUser();
+      }
+    } catch (error) {
+      // Silent fail; the user can still log in manually.
+    }
   };
 
   //
@@ -80,7 +136,11 @@ AuthenticationRequiredModalContentConnector.propTypes = {
   dispatchFetchGeneralSettings: PropTypes.func.isRequired,
   dispatchSetGeneralSettingsValue: PropTypes.func.isRequired,
   dispatchSaveGeneralSettings: PropTypes.func.isRequired,
-  dispatchFetchStatus: PropTypes.func.isRequired
+  dispatchFetchStatus: PropTypes.func.isRequired,
+  dispatchFetchCurrentUser: PropTypes.func.isRequired,
+  isSaving: PropTypes.bool,
+  saveError: PropTypes.object,
+  settings: PropTypes.object
 };
 
 export default connect(createMapStateToProps, mapDispatchToProps)(AuthenticationRequiredModalContentConnector);